Add support for loading items from chunk data
This commit includes a change of terminology from 'spawn' to 'entity' as
the noun for an item contained within a chunk. This matches the name
used in the NBT data and is more consistent with the types of objects
that are referenced in that data.

The verb 'spawn' still refers to creating a new object in the game
world, so you still 'spawn' an 'entity' in a chunk, as expected.

There are two outstanding items for handling non-item entities and for
loading the motion/velocity data (which I was not able to get working).