Convert repository to use gb for building
gb is a tool that is similar to godag, but much more frequently updated
and capable of building scripts that allow people without gb to build
the source repository. This removes a dependency for people installing
chunkymonkey which is a good thing.

Each directory has a Makefile which can be run directly, although they
will need to be invoked in the appropriate order. Running the build.sh
script in the src directory will perform these builds incrementally for
anyone without 'gb'. To build it with gb, simply run the gb command
within the src directory.